requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
pandas==2.3.3
numpy==2.2.6
faiss-cpu==1.13.2
sentence-transformers==5.4.1
rank-bm25==0.2.2
groq==1.1.2
anthropic==0.96.0
